Rename slice::unzip() to vec::unzip()
unzip() has nothing to do with slices, so it belongs in vec.
